Trash screen added
Adding trash screen and handle logics from moving email to trash, to choose an email in trash screen for permanently delete or recover it. Additionally, some of the screens' titles are synchronized with the app's title, as well as a field called: "isTrashed" is created to mark if that email is putted to trash.